backoff for try_write on windows
